feat: change price calculates and validations

+++ DemoTiendasApi/Models/PurchaseProduct.cs
@@ -3,12 +3,12 @@
     public class PurchaseProduct
     {
         public int ProductId { get; set; }
-        public float CostWithoutTax { get; set; }
-        public float PurchaseTax { get; set; }
-        public float CostWithTax { get; set; }
+        public decimal CostWithoutTax { get; set; }
+        public decimal PurchaseTax { get; set; }
+        public decimal CostWithTax { get; set; }
         public int Quantity { get; set; }
-        public float ProfitPercentage { get; set; } //porcentaje de ganancia
-        public float SalePrice { get; set; }
+        public decimal ProfitPercentage { get; set; } //porcentaje de ganancia
+        public decimal SalePrice { get; set; }
         public decimal TotalPrice { get; set; }
     }
 }
